requirements.txt
Flask==2.3.2
psycopg2-binary==2.9.6
pydantic==1.10.8
requests==2.31.0
SQLAlchemy==1.4.48
Werkzeug==2.3.4
python-dotenv